Update Robot.php
Changed the pass method to account for defense from other robots, essentially it just lowers the number that the "check" must roll below. I also changed getDefended to return a fraction of the maxDefense score. I am thinking that something better might be in order, something that accounts for the standard deviation from the average defense.